% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleaning raw NOAA data frame and returning a clean data frame}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{A raw data frame containing eqrthquake data from NOAA website}
}
\value{
A clean data frame with cleaned date, latitude,longitude and death columns
}
\description{
Cleaning raw NOAA data frame and returning a clean data frame
}
\details{
In order to use this function, one has to obtain the NOAA data from
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}. The function returns
a cleaned data frame with cleaned columns: \code{date,LONGITUDE,LATITUDE,DEATHS}. The \code{date} column is
converted into date-format and the \code{LONGITUDE,LATITUDE,DEATHS} are converted into numeric types.
The function utilizes the function \code{eq_location_clean}.
}
\examples{
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\\t")
data <- eq_clean_data(data)
}

}
