% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create Pop-Ups to be utilized by the function eq_map()}
\usage{
eq_create_label(df_map)
}
\arguments{
\item{df_map}{data frame. cleaned NOAA data}
}
\value{
Returns a string/HTML label with the right format for the columns \code{LOCATION_NAME,EQ_PRIMARY,TOTAL_DEATHS} as an annotation text in the leaflet map.
}
\description{
Create Pop-Ups to be utilized by the function eq_map()
}
\details{
Takes the dataset as an argument and creates an HTML label that can be used as the annotation text in the leaflet map.
}
\examples{
\dontrun{
library(magrittr)
readr::read_delim("signif.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}
}
