% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning raw NOAA data frame location columns}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{A raw data frame containing eqrthquake data from NOAA website}
}
\value{
A clean data frame with location column
}
\description{
Cleaning raw NOAA data frame location columns
}
\details{
In order to use this function, one has to obtain the NOAA data from
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}. The function returns
a cleaned data frame with cleaned columns: \code{LOCATION_NAME}. From the \code{LOCATION_NAME}
column the country name is stripped out and names are converted to title case.
}
\examples{
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\\t")
data <- eq_location_clean(data)
}

}
