% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plotting a timeline of earthquakes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aesthetic mappings}

\item{data}{data to be visualized by the geom}

\item{stat}{character. The statistical transformation to use on the data for this layer, as a string}

\item{position}{character. Position adjustment, either as a string, or the result of a call to a position adjustment function}

\item{na.rm}{logical. variable to indicate if NAs should be removed from the data}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{logical. If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification.}

\item{...}{Other arguments passed on to the layer}
}
\value{
Returns a layer, a combination of data, stat and geom objects.
}
\description{
Plotting a timeline of earthquakes
}
\examples{
#Before running the function on the data,
#one needs to read in the NOAA data and clean the data frame in order
#for the function to work properly.
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\\t")
data <- eq_clean_data(data)
 data \%>\%
 dplyr::filter(COUNTRY == c("MEXICO","China") & lubridate::year(DATE) >= 2010) \%>\%
 ggplot(aes(x=DATE,y=COUNTRY,color=TOTAL_DEATHS,size=EQ_PRIMARY, label=LOCATION_NAME)) +
 geom_timeline(alpha=0.7) +
 theme(legend.position="bottom", legend.box="horizontal", plot.title=element_text(hjust=0.5)) +
 ggtitle("Earthquakes Timeline") +
 labs(size = "Richter scale value", color = "# Deaths")
}

}
