% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom: Timeline for NOAA data}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
GeomTimeline represents the source code for the geom \code{TimeLine} without labels.
It corresponds with \code{geom_timeline} and represents the geom object that is returned as part of the layer.
}
\keyword{datasets}
