% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimelinelabel}
\alias{GeomTimelinelabel}
\title{Geom: Adding labels to the Timeline-Geom}
\format{An object of class \code{GeomTimelinelabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelinelabel
}
\description{
GeomTimeline represents the source code for the geom \code{TimeLine} with labels.
It corresponds with \code{geom_timeline} and adds label objects to the timeline layer.
}
\keyword{datasets}
